+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## v0.0.4
+- Add View of Custom Game Documents
+
 ## v0.0.3
 - Add View of Panorama CSS
 - Add View of Abilities Data Driven